Trabalho de programação finalizado (versão 1.0).

+++ com/github/stefaniojr/prog3/project/domain/atividades/Avaliacao.java
@@ -7,23 +7,23 @@ import com.github.stefaniojr.prog3.project.domain.Estudante;
 public class Avaliacao implements Serializable {
     private static final long serialVersionUID = 1348633635465464569L;
     private Estudante estudante;
-    private float nota;
+    private double nota;
 
-    public Avaliacao(Estudante estudante, float nota){
+    public Avaliacao(Estudante estudante, double nota) {
         this.estudante = estudante;
         this.nota = nota;
-      }
-    
+    }
+
     // Getters.
-    public Estudante obterAvaliador(){
+    public Estudante obterAvaliador() {
         return estudante;
     }
 
-    public Estudante obterRef(){
+    public Estudante obterRef() {
         return obterAvaliador();
     }
 
-    public float obterNota(){
+    public double obterNota() {
         return nota;
     }
 }
